+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 ## Azure NetAppFiles Cross-Region Replication (CRR) Sample Script PowerShell
 
+[1.0.1 (2021-08-30)](#1.0.1 (2021-08-30))
 [1.0.0 (2021-01-20)](#1.0.0 (2021-01-20))
 
-# 1.0.0 (2019-10-09)
+# 1.0.1 (2021-08-03)
+
+*Features*
+* Initial commit
+
+*Bug Fixes*
+* Fix authorize functionality by adding VolumeType property to data replication volume
+* Add WaitForMirrorStatus after authorizing.
+
+*Breaking Changes*
+* N/A
+
+# 1.0.0 (2021-01-20)
 
 *Features*
 * Initial commit